ISIS-1976: moving responsibilities from PersistenceSession to ObjectAdapterProvider
introduces ObjectAdapterProvider.Delegating
removes AdapterManager
ports changes from DN5-plugin to DN-4 plugin

Task-Url: https://issues.apache.org/jira/browse/ISIS-1976
